separate AddUnreportedExpenseFooter
