+++ src/pages/AddUnreportedExpense.tsx
@@ -1,12 +1,9 @@
 import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
-import {InteractionManager} from 'react-native';
 import type {OnyxCollection} from 'react-native-onyx';
-import Button from '@components/Button';
+import AddUnreportedExpenseFooter from '@components/AddUnreportedExpenseFooter';
 import EmptyStateComponent from '@components/EmptyStateComponent';
-import FormHelpMessage from '@components/FormHelpMessage';
 import HeaderWithBackButton from '@components/HeaderWithBackButton';
 import LottieAnimations from '@components/LottieAnimations';
-import {useSession} from '@components/OnyxListItemProvider';
 import ScreenWrapper from '@components/ScreenWrapper';
 import SelectionList from '@components/SelectionList';
 import type {ListItem, SelectionListHandle} from '@components/SelectionList/types';
@@ -15,7 +12,6 @@ import useDebouncedState from '@hooks/useDebouncedState';
 import useLocalize from '@hooks/useLocalize';
 import useNetwork from '@hooks/useNetwork';
 import useOnyx from '@hooks/useOnyx';
-import usePermissions from '@hooks/usePermissions';
 import usePolicy from '@hooks/usePolicy';
 import useThemeStyles from '@hooks/useThemeStyles';
 import {fetchUnreportedExpenses} from '@libs/actions/UnreportedExpenses';
@@ -24,14 +20,13 @@ import getNonEmptyStringOnyxID from '@libs/getNonEmptyStringOnyxID';
 import interceptAnonymousUser from '@libs/interceptAnonymousUser';
 import type {AddUnreportedExpensesParamList} from '@libs/Navigation/types';
 import {canSubmitPerDiemExpenseFromWorkspace, getPerDiemCustomUnit} from '@libs/PolicyUtils';
-import {getTransactionDetails, isIOUReport} from '@libs/ReportUtils';
+import {getTransactionDetails} from '@libs/ReportUtils';
 import {shouldRestrictUserBillableActions} from '@libs/SubscriptionUtils';
 import tokenizedSearch from '@libs/tokenizedSearch';
 import {createUnreportedExpenses, getAmount, getCurrency, getDescription, getMerchant, isPerDiemRequest} from '@libs/TransactionUtils';
 import Navigation from '@navigation/Navigation';
 import type {PlatformStackScreenProps} from '@navigation/PlatformStackNavigation/types';
-import {convertBulkTrackedExpensesToIOU, startMoneyRequest} from '@userActions/IOU';
-import {changeTransactionsReport} from '@userActions/Transaction';
+import {startMoneyRequest} from '@userActions/IOU';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
@@ -57,14 +52,7 @@ function AddUnreportedExpense({route}: AddUnreportedExpensePageType) {
     const [policyCategories] = useOnyx(`${ONYXKEYS.COLLECTION.POLICY_CATEGORIES}${getNonEmptyStringOnyxID(report?.policyID)}`, {canBeMissing: true});
     const [hasMoreUnreportedTransactionsResults] = useOnyx(ONYXKEYS.HAS_MORE_UNREPORTED_TRANSACTIONS_RESULTS, {canBeMissing: true});
     const [isLoadingUnreportedTransactions] = useOnyx(ONYXKEYS.IS_LOADING_UNREPORTED_TRANSACTIONS, {canBeMissing: true});
-    const {isBetaEnabled} = usePermissions();
-    const isASAPSubmitBetaEnabled = isBetaEnabled(CONST.BETAS.ASAP_SUBMIT);
-    const [policyRecentlyUsedCurrencies] = useOnyx(ONYXKEYS.RECENTLY_USED_CURRENCIES, {canBeMissing: true});
-    const [quickAction] = useOnyx(ONYXKEYS.NVP_QUICK_ACTION_GLOBAL_CREATE, {canBeMissing: true});
-    const session = useSession();
-    const [transactionViolations] = useOnyx(ONYXKEYS.COLLECTION.TRANSACTION_VIOLATIONS, {canBeMissing: true});
     const shouldShowUnreportedTransactionsSkeletons = isLoadingUnreportedTransactions && hasMoreUnreportedTransactionsResults && !isOffline;
-    const [allTransactions] = useOnyx(ONYXKEYS.COLLECTION.TRANSACTION, {canBeMissing: true});
 
     const getUnreportedTransactions = useCallback(
         (transactions: OnyxCollection<Transaction>) => {
@@ -161,80 +149,6 @@ function AddUnreportedExpense({route}: AddUnreportedExpensePageType) {
             isSelected: selectedIds.has(item.transactionID),
         }));
     }, [filteredTransactions, selectedIds]);
-
-    const handleConfirm = useCallback(() => {
-        if (selectedIds.size === 0) {
-            setErrorMessage(translate('iou.selectUnreportedExpense'));
-            return;
-        }
-        Navigation.dismissModal();
-        // eslint-disable-next-line @typescript-eslint/no-deprecated
-        InteractionManager.runAfterInteractions(() => {
-            if (report && isIOUReport(report)) {
-                convertBulkTrackedExpensesToIOU(
-                    [...selectedIds],
-                    report.reportID,
-                    isASAPSubmitBetaEnabled,
-                    session?.accountID ?? CONST.DEFAULT_NUMBER_ID,
-                    session?.email ?? '',
-                    transactionViolations,
-                    policyRecentlyUsedCurrencies ?? [],
-                    quickAction,
-                );
-            } else {
-                changeTransactionsReport({
-                    transactionIDs: [...selectedIds],
-                    isASAPSubmitBetaEnabled,
-                    accountID: session?.accountID ?? CONST.DEFAULT_NUMBER_ID,
-                    email: session?.email ?? '',
-                    newReport: reportToConfirm,
-                    policy,
-                    reportNextStep,
-                    policyCategories,
-                    allTransactionsCollection: allTransactions,
-                });
-            }
-        });
-        setErrorMessage('');
-    }, [
-        selectedIds,
-        translate,
-        report,
-        isASAPSubmitBetaEnabled,
-        session?.accountID,
-        session?.email,
-        transactionViolations,
-        policyRecentlyUsedCurrencies,
-        quickAction,
-        reportToConfirm,
-        policy,
-        reportNextStep,
-        policyCategories,
-        allTransactions,
-    ]);
-
-    const footerContent = useMemo(() => {
-        return (
-            <>
-                {!!errorMessage && (
-                    <FormHelpMessage
-                        style={[styles.ph1, styles.mb2]}
-                        isError
-                        message={errorMessage}
-                    />
-                )}
-                <Button
-                    success
-                    large
-                    style={[styles.w100, styles.justifyContentCenter]}
-                    text={translate('iou.addUnreportedExpenseConfirm')}
-                    onPress={handleConfirm}
-                    pressOnEnter
-                    enterKeyEventListenerPriority={1}
-                />
-            </>
-        );
-    }, [errorMessage, styles, translate, handleConfirm]);
 
     const headerMessage = useMemo(() => {
         if (debouncedSearchValue.trim() && unreportedExpenses?.length === 0) {
@@ -362,7 +276,18 @@ function AddUnreportedExpense({route}: AddUnreportedExpensePageType) {
                 onEndReachedThreshold={0.75}
                 addBottomSafeAreaPadding
                 listFooterContent={shouldShowUnreportedTransactionsSkeletons ? <UnreportedExpensesSkeleton fixedNumberOfItems={3} /> : undefined}
-                footerContent={footerContent}
+                footerContent={
+                    <AddUnreportedExpenseFooter
+                        selectedIds={selectedIds}
+                        report={report}
+                        reportToConfirm={reportToConfirm}
+                        reportNextStep={reportNextStep}
+                        policy={policy}
+                        policyCategories={policyCategories}
+                        errorMessage={errorMessage}
+                        setErrorMessage={setErrorMessage}
+                    />
+                }
                 disableMaintainingScrollPosition
             />
         </ScreenWrapper>
